Enable HMR
